requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
pydantic>=2.7.0
werkzeug>=3.0.0
cryptography>=42.0.0
celery[redis]>=5.4.0
redis>=5.0.0
requests>=2.31.0
httpx>=0.27.0
mysql-connector-python>=9.0.0
faker>=25.0.0
bitarray>=2.9.0
mmh3>=4.1.0
starlette>=0.37.0
itsdangerous>=2.2.0
